// src/enums/AttributeType.ts
export enum AttributeType {
  NUMBER = "number",
  STRING = "string",
  BOOLEAN = "boolean",
  DATE = "date",
  MAP = "map",
  LIST = "list",
  STRING_LIST = "list<string>",
  NUMBER_LIST = "list<number>",
  BOOLEAN_LIST = "list<boolean>",
  DATE_LIST = "list<date>",
  MAP_LIST = "list<map>",
  STRING_SET = "set<string>",
  NUMBER_SET = "set<number>",
}